UserGoal Controller Tests. I had to edit some of goal backend stuff because it wasn't the same as what was in the Goal model/dto.
However this shouldn't cause any issues because these methods weren't actually being used so we should be good. Still might be worth a check looking into goals just to make sure.

+++ backend/src/main/java/recipes/fridger/backend/service/GoalService.java
@@ -15,6 +15,6 @@ public interface GoalService {
     public Goal getGoalByID(Long id);
 
     //get all goals matching params
-    public Iterable<Goal> getGoals(Long id, String endGoal, Double calories, Double carbs, Double protein, Double fat, Double currWeight, Double goalWeight, Long userId);
+    public Iterable<Goal> getGoals(Long id, String endGoal, Integer calories, Integer carbs, Integer protein, Integer fat, Double currWeight, Double goalWeight, Long userId);
 
 }
