+++ backend/src/main/java/recipes/fridger/backend/service/GoalServiceImpl.java
@@ -54,7 +54,7 @@ public class GoalServiceImpl implements GoalService{
     //get goals matching params
     @Transactional
     @Override
-    public Iterable<Goal> getGoals(Long id, String endGoal, Double calories, Double carbs, Double protein, Double fat, Double currWeight, Double goalWeight, Long userId) {
+    public Iterable<Goal> getGoals(Long id, String endGoal, Integer calories, Integer carbs, Integer protein, Integer fat, Double currWeight, Double goalWeight, Long userId) {
         return goals.find(id, endGoal, calories, carbs, protein, fat, currWeight, goalWeight, userId);
     }
 }
